Revise preset theme JSON structure
- Move colors in theme JSON into a colors property, add images - better
  matches the internal structure in Redux store

- Also reorganize utility functions related to themes

Fixes #113
