+++ src/preset-themes/001.json
@@ -1,35 +1,40 @@
 {
   "title": "001",
-  "toolbar": {
-    "h": 187,
-    "s": 77,
-    "l": 53,
-    "a": 100
+  "colors": {
+    "toolbar": {
+      "h": 187,
+      "s": 77,
+      "l": 53,
+      "a": 100
+    },
+    "toolbar_text": {
+      "h": 213,
+      "s": 43,
+      "l": 89
+    },
+    "accentcolor": {
+      "h": 187,
+      "s": 77,
+      "l": 53
+    },
+    "textcolor": {
+      "h": 187,
+      "s": 0,
+      "l": 33
+    },
+    "toolbar_field": {
+      "h": 211,
+      "s": 0,
+      "l": 100,
+      "a": 53.33333206176758
+    },
+    "toolbar_field_text": {
+      "h": 317,
+      "s": 100,
+      "l": 64
+    }
   },
-  "toolbar_text": {
-    "h": 213,
-    "s": 43,
-    "l": 89
-  },
-  "accentcolor": {
-    "h": 187,
-    "s": 77,
-    "l": 53
-  },
-  "textcolor": {
-    "h": 187,
-    "s": 0,
-    "l": 33
-  },
-  "toolbar_field": {
-    "h": 211,
-    "s": 0,
-    "l": 100,
-    "a": 53.33333206176758
-  },
-  "toolbar_field_text": {
-    "h": 317,
-    "s": 100,
-    "l": 64
+  "images": {
+    "headerURL": ""
   }
 }
