+++ src/preset-themes/002.json
@@ -1,35 +1,40 @@
 {
   "title": "002",
-  "toolbar": {
-    "h": 4,
-    "s": 0,
-    "l": 100,
-    "a": 100
+  "colors": {
+    "toolbar": {
+      "h": 4,
+      "s": 0,
+      "l": 100,
+      "a": 100
+    },
+    "toolbar_text": {
+      "h": 172,
+      "s": 0,
+      "l": 8
+    },
+    "accentcolor": {
+      "h": 266,
+      "s": 35,
+      "l": 26
+    },
+    "textcolor": {
+      "h": 174,
+      "s": 86,
+      "l": 50
+    },
+    "toolbar_field": {
+      "h": 260,
+      "s": 74,
+      "l": 51,
+      "a": 11
+    },
+    "toolbar_field_text": {
+      "h": 172,
+      "s": 0,
+      "l": 3
+    }
   },
-  "toolbar_text": {
-    "h": 172,
-    "s": 0,
-    "l": 8
-  },
-  "accentcolor": {
-    "h": 266,
-    "s": 35,
-    "l": 26
-  },
-  "textcolor": {
-    "h": 174,
-    "s": 86,
-    "l": 50
-  },
-  "toolbar_field": {
-    "h": 260,
-    "s": 74,
-    "l": 51,
-    "a": 11
-  },
-  "toolbar_field_text": {
-    "h": 172,
-    "s": 0,
-    "l": 3
+  "images": {
+    "headerURL": ""
   }
 }
