+++ src/web/lib/components/AppBackground/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { colorToCSS } from "../../../../lib/utils";
+import { colorToCSS } from "../../../../lib/themes";
 
 import "./index.scss";
