+++ src/web/lib/components/ThemeBackgroundPicker/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import classnames from "classnames";
 import onClickOutside from "react-onclickoutside";
 
-import { colorToCSS } from "../../../../lib/utils";
+import { colorToCSS } from "../../../../lib/themes";
 import Metrics from "../../../../lib/metrics";
 
 import "./index.scss";
