+++ src/web/lib/components/ThemeColorsEditor/index.js
@@ -3,7 +3,7 @@ import classnames from "classnames";
 import { SketchPicker } from "react-color";
 import onClickOutside from "react-onclickoutside";
 import { colorLabels, colorsWithAlpha } from "../../../../lib/constants";
-import { colorToCSS } from "../../../../lib/utils";
+import { colorToCSS } from "../../../../lib/themes";
 import Metrics from "../../../../lib/metrics";
 
 import "./index.scss";
